+++ library/src/main/java/com/getstream/sdk/chat/adapter/AttachmentListItemAdapter.java
@@ -64,18 +64,17 @@ public class AttachmentListItemAdapter extends RecyclerView.Adapter<RecyclerView
     @Override
     public void onBindViewHolder(RecyclerView.ViewHolder holder, int position) {
         Attachment attachment = attachments.get(position);
-        ((BaseAttachmentViewHolder) holder).setBubbleHelper(bubbleHelper);
-        ((BaseAttachmentViewHolder) holder).bind(this.context,
+        ((BaseAttachmentViewHolder) holder).bind(
+                context,
                 messageListItem,
+                message,
                 attachment,
                 style,
+                bubbleHelper,
                 attachmentClickListener,
                 longClickListener);
     }
 
-    public MessageListViewStyle getStyle() {
-        return style;
-    }
 
     public void setStyle(MessageListViewStyle style) {
         this.style = style;
@@ -95,10 +94,6 @@ public class AttachmentListItemAdapter extends RecyclerView.Adapter<RecyclerView
 
     public void setGiphySendListener(MessageListView.GiphySendListener giphySendListener) {
         this.giphySendListener = giphySendListener;
-    }
-
-    public MessageListView.BubbleHelper getBubbleHelper() {
-        return bubbleHelper;
     }
 
     public void setBubbleHelper(MessageListView.BubbleHelper bubbleHelper) {