+++ library/src/main/java/com/getstream/sdk/chat/adapter/AttachmentViewHolder.java
@@ -10,6 +10,8 @@ import android.widget.AdapterView;
 import android.widget.ListView;
 import android.widget.TextView;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import androidx.constraintlayout.widget.ConstraintLayout;
 
 import com.bumptech.glide.Glide;
@@ -17,6 +19,7 @@ import com.getstream.sdk.chat.R;
 import com.getstream.sdk.chat.StreamChat;
 import com.getstream.sdk.chat.model.Attachment;
 import com.getstream.sdk.chat.model.ModelType;
+import com.getstream.sdk.chat.rest.Message;
 import com.getstream.sdk.chat.utils.roundedImageView.PorterShapeImageView;
 import com.getstream.sdk.chat.view.MessageListView;
 import com.getstream.sdk.chat.view.MessageListViewStyle;
@@ -28,15 +31,25 @@ public class AttachmentViewHolder extends BaseAttachmentViewHolder {
 
     final String TAG = AttachmentViewHolder.class.getSimpleName();
     // Attachment
-    private ConstraintLayout cl_attachment_media;
+    private ConstraintLayout cl_attachment_media, cl_des;
     private PorterShapeImageView iv_media_thumb;
     private ListView lv_attachment_file;
     private TextView tv_media_title, tv_media_play, tv_media_des;
+
+    private Context context;
+    private MessageListItem messageListItem;
+    private Message message;
+    private Attachment attachment;
+    private MessageListViewStyle style;
+    private MessageListView.BubbleHelper bubbleHelper;
+    private MessageListView.AttachmentClickListener clickListener;
+    private MessageListView.MessageLongClickListener longClickListener;
 
     public AttachmentViewHolder(int resId, ViewGroup parent) {
         super(resId, parent);
         // Attach
         cl_attachment_media = itemView.findViewById(R.id.cl_attachment_media);
+        cl_des = itemView.findViewById(R.id.cl_des);
         iv_media_thumb = itemView.findViewById(R.id.iv_media_thumb);
         lv_attachment_file = itemView.findViewById(R.id.lv_attachment_file);
         tv_media_title = itemView.findViewById(R.id.tv_media_title);
@@ -45,14 +58,24 @@ public class AttachmentViewHolder extends BaseAttachmentViewHolder {
     }
 
     @Override
-    public void bind(Context context,
-                     MessageListItem messageListItem,
-                     Attachment attachment,
-                     MessageListViewStyle style,
-                     MessageListView.AttachmentClickListener clickListener,
-                     MessageListView.MessageLongClickListener longClickListener) {
+    public void bind(@NonNull Context context,
+                     @NonNull MessageListItem messageListItem,
+                     @NonNull Message message,
+                     @NonNull Attachment attachment,
+                     @NonNull MessageListViewStyle style,
+                     @NonNull MessageListView.BubbleHelper bubbleHelper,
+                     @Nullable MessageListView.AttachmentClickListener clickListener,
+                     @Nullable MessageListView.MessageLongClickListener longClickListener) {
 
-        super.bind(context, messageListItem, attachment, style, clickListener, longClickListener);
+        this.context = context;
+        this.messageListItem = messageListItem;
+        this.message = message;
+        this.attachment = attachment;
+        this.style = style;
+        this.bubbleHelper = bubbleHelper;
+        this.clickListener = clickListener;
+        this.longClickListener = longClickListener;
+
         applyStyle();
         configAttachment();
     }
@@ -103,13 +126,17 @@ public class AttachmentViewHolder extends BaseAttachmentViewHolder {
     }
 
     private void configImageThumbBackground() {
-        Drawable background = getBubbleHelper().getDrawableForAttachment(message, getMessageListItem().isMine(), getMessageListItem().getPositions(), attachment);
+        Drawable background = bubbleHelper.getDrawableForAttachment(message, messageListItem.isMine(), messageListItem.getPositions(), attachment);
         iv_media_thumb.setShape(context, background);
+        if (tv_media_des.getVisibility() == View.VISIBLE || tv_media_title.getVisibility() == View.VISIBLE) {
+            background = bubbleHelper.getDrawableForAttachmentDescription(messageListItem.getMessage(), messageListItem.isMine(), messageListItem.getPositions());
+            cl_des.setBackground(background);
+        }
     }
 
     private void configAttachViewBackground(View view) {
-        if (style.getMessageBubbleDrawable(getMessageListItem().isMine()) != -1)
-            view.setBackground(context.getDrawable(style.getMessageBubbleDrawable(getMessageListItem().isMine())));
+        if (style.getMessageBubbleDrawable(messageListItem.isMine()) != -1)
+            view.setBackground(context.getDrawable(style.getMessageBubbleDrawable(messageListItem.isMine())));
     }
 
 
@@ -128,12 +155,13 @@ public class AttachmentViewHolder extends BaseAttachmentViewHolder {
         lv_attachment_file.setOnItemClickListener((AdapterView<?> parent, View view,
                                                    int position, long id) -> {
             Log.d(TAG, "Attach onMessageClick: " + position);
-            triggerClick();
-
-
+            if (clickListener != null)
+                clickListener.onAttachmentClick(message, attachment);
         });
+
         lv_attachment_file.setOnItemLongClickListener((AdapterView<?> parent, View view, int position_, long id) -> {
-            triggerLongClick();
+            if (longClickListener != null)
+                longClickListener.onMessageLongClick(message);
             return true;
         });
 
@@ -171,15 +199,8 @@ public class AttachmentViewHolder extends BaseAttachmentViewHolder {
         cl_attachment_media.setVisibility(View.VISIBLE);
         configAttachViewBackground(cl_attachment_media);
         configImageThumbBackground();
+        configClickListeners();
 
-        // Set Click Listener
-        cl_attachment_media.setOnClickListener(this);
-
-
-        cl_attachment_media.setOnLongClickListener((View v) -> {
-            this.triggerLongClick();
-            return true;
-        });
         if (!attachUrl.contains("https:"))
             attachUrl = "https:" + attachUrl;
         Glide.with(context)
@@ -205,27 +226,26 @@ public class AttachmentViewHolder extends BaseAttachmentViewHolder {
             tv_media_play.setVisibility(View.GONE);
     }
 
-    private void triggerLongClick() {
-        if (this.getLongClickListener() != null) {
-            this.getLongClickListener().onMessageLongClick(message);
-        }
-    }
-
-    private void triggerClick() {
-        if (this.getClickListener() != null) {
-            this.getClickListener().onAttachmentClick(message, attachment);
-        }
-    }
-
     private void applyStyle() {
-
-        if(getMessageListItem().isMine()) {
+        if(messageListItem.isMine()) {
             style.attachmentTitleTextMine.apply(tv_media_title);
             style.attachmentDescriptionTextMine.apply(tv_media_des);
         } else {
             style.attachmentTitleTextTheirs.apply(tv_media_title);
             style.attachmentDescriptionTextTheirs.apply(tv_media_des);
         }
+    }
 
+    private void configClickListeners(){
+        cl_attachment_media.setOnClickListener(view -> {
+            if (clickListener != null)
+                clickListener.onAttachmentClick(message, attachment);
+        });
+
+        cl_attachment_media.setOnLongClickListener(view -> {
+            if (longClickListener != null)
+                longClickListener.onMessageLongClick(message);
+            return true;
+        });
     }
 }