+++ library/src/main/java/com/getstream/sdk/chat/adapter/AttachmentViewHolderFile.java
@@ -7,10 +7,13 @@ import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import androidx.constraintlayout.widget.ConstraintLayout;
 
 import com.getstream.sdk.chat.R;
 import com.getstream.sdk.chat.model.Attachment;
+import com.getstream.sdk.chat.rest.Message;
 import com.getstream.sdk.chat.view.MessageListView;
 import com.getstream.sdk.chat.view.MessageListViewStyle;
 
@@ -22,6 +25,14 @@ public class AttachmentViewHolderFile extends BaseAttachmentViewHolder {
     private ImageView iv_file_thumb;
     private TextView tv_file_size, tv_file_title;
 
+    private Context context;
+    private MessageListItem messageListItem;
+    private Message message;
+    private Attachment attachment;
+    private MessageListViewStyle style;
+    private MessageListView.BubbleHelper bubbleHelper;
+    private MessageListView.AttachmentClickListener clickListener;
+    private MessageListView.MessageLongClickListener longClickListener;
 
     public AttachmentViewHolderFile(int resId, ViewGroup parent) {
         super(resId, parent);
@@ -33,20 +44,31 @@ public class AttachmentViewHolderFile extends BaseAttachmentViewHolder {
     }
 
     @Override
-    public void bind(Context context,
-                     MessageListItem messageListItem,
-                     Attachment attachment,
-                     MessageListViewStyle style,
-                     MessageListView.AttachmentClickListener clickListener,
-                     MessageListView.MessageLongClickListener longClickListener) {
-        super.bind(context, messageListItem, attachment, style, clickListener, longClickListener);
+    public void bind(@NonNull Context context,
+                     @NonNull MessageListItem messageListItem,
+                     @NonNull Message message,
+                     @NonNull Attachment attachment,
+                     @NonNull MessageListViewStyle style,
+                     @NonNull MessageListView.BubbleHelper bubbleHelper,
+                     @Nullable MessageListView.AttachmentClickListener clickListener,
+                     @Nullable MessageListView.MessageLongClickListener longClickListener) {
+
+        this.context = context;
+        this.messageListItem = messageListItem;
+        this.message = message;
+        this.attachment = attachment;
+        this.style = style;
+        this.bubbleHelper = bubbleHelper;
+        this.clickListener = clickListener;
+        this.longClickListener = longClickListener;
+
         applyStyle();
         configAttachment();
+        configClickListeners();
     }
 
     private void applyStyle() {
-
-        if(getMessageListItem().isMine()) {
+        if(messageListItem.isMine()) {
             style.attachmentTitleTextMine.apply(tv_file_title);
             style.attachmentFileSizeTextMine.apply(tv_file_size);
         } else {
@@ -61,14 +83,25 @@ public class AttachmentViewHolderFile extends BaseAttachmentViewHolder {
         iv_file_thumb.setImageResource(attachment.getIcon());
         tv_file_title.setText(attachment.getTitle());
 
-        Drawable background = getBubbleHelper().getDrawableForAttachment(getMessageListItem().getMessage(),
-                getMessageListItem().isMine(),
-                getMessageListItem().getPositions(),
+        Drawable background = bubbleHelper.getDrawableForAttachment(messageListItem.getMessage(),
+                messageListItem.isMine(),
+                messageListItem.getPositions(),
                 attachment);
         cl_attachment.setBackground(background);
 
-        cl_attachment.setOnClickListener(this);
-        cl_attachment.setOnLongClickListener(this);
+    }
+
+    private void configClickListeners(){
+        cl_attachment.setOnClickListener(view -> {
+            if (clickListener != null)
+                clickListener.onAttachmentClick(message, attachment);
+        });
+
+        cl_attachment.setOnLongClickListener(view -> {
+            if (longClickListener != null)
+                longClickListener.onMessageLongClick(message);
+            return true;
+        });
     }
 }