+++ library/src/main/java/com/getstream/sdk/chat/adapter/ThreadSeparatorViewHolder.java
@@ -4,29 +4,29 @@ import android.content.Context;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
+import androidx.annotation.NonNull;
+
 import com.getstream.sdk.chat.R;
 import com.getstream.sdk.chat.rest.response.ChannelState;
+import com.getstream.sdk.chat.view.MessageListView;
 import com.getstream.sdk.chat.view.MessageListViewStyle;
 
 public class ThreadSeparatorViewHolder extends BaseMessageListItemViewHolder {
-    private MessageListViewStyle style;
-    private TextView tv_text;
 
+    private TextView tv_text;
     public ThreadSeparatorViewHolder(int resId, ViewGroup viewGroup) {
         super(resId, viewGroup);
         tv_text = itemView.findViewById(R.id.tv_text);
     }
 
     @Override
-    public void bind(Context context,
-                     ChannelState channelState,
-                     MessageListItem messageListItem,
+    public void bind(@NonNull Context context,
+                     @NonNull ChannelState channelState,
+                     @NonNull MessageListItem messageListItem,
+                     @NonNull MessageListViewStyle style,
+                     @NonNull MessageListView.BubbleHelper bubbleHelper,
+                     @NonNull MessageViewHolderFactory factory,
                      int position) {
 //        tv_text.setText();
-    }
-
-    @Override
-    public void setStyle(MessageListViewStyle style) {
-        this.style = style;
     }
 }