+++ library/src/main/java/com/getstream/sdk/chat/viewmodel/ChannelListViewModel.java
@@ -19,9 +19,11 @@ import com.getstream.sdk.chat.model.Channel;
 import com.getstream.sdk.chat.model.Event;
 import com.getstream.sdk.chat.rest.core.ChatEventHandler;
 import com.getstream.sdk.chat.rest.core.Client;
+import com.getstream.sdk.chat.rest.interfaces.ChannelCallback;
 import com.getstream.sdk.chat.rest.interfaces.CompletableCallback;
 import com.getstream.sdk.chat.rest.interfaces.QueryChannelListCallback;
 import com.getstream.sdk.chat.rest.request.QueryChannelsRequest;
+import com.getstream.sdk.chat.rest.response.ChannelResponse;
 import com.getstream.sdk.chat.rest.response.ChannelState;
 import com.getstream.sdk.chat.rest.response.CompletableResponse;
 import com.getstream.sdk.chat.rest.response.QueryChannelsResponse;
@@ -449,7 +451,7 @@ public class ChannelListViewModel extends AndroidViewModel implements LifecycleH
         channels.postValue(channelCopy);
     }
 
-    private boolean deleteChannel(Channel channel) {
+    public boolean deleteChannel(Channel channel) {
         List<Channel> channelCopy = channels.getValue();
         if (channelCopy == null) {
             channelCopy = new ArrayList<>();
@@ -476,6 +478,7 @@ public class ChannelListViewModel extends AndroidViewModel implements LifecycleH
         channelCopy.addAll(newChannels);
         channels.postValue(channelCopy);
     }
+
 
     private void queryChannelsInner(int attempt) {